pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "uk-tax-calculator"
version = "0.0.0"
description = "A calculator for UK tax."
authors = ["Bilbottom"]
readme = "README.md"
packages = [{include = "uk_tax_calculator"}]

[tool.poetry.dependencies]
python = "^3.11"
duckdb = "^1.1.0"

[tool.poetry.group]
dev.optional = true
test.optional = true

[tool.poetry.group.dev.dependencies]
coverage-badge = "^1.1.0"
pre-commit = "^3.6.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
pyyaml = "^6.0.2"


[tool.pytest.ini_options]
addopts = "--cov=uk_tax_calculator --cov-fail-under=80"
testpaths = ["tests"]


[tool.ruff]
line-length = 80
indent-width = 4
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["F", "I", "N", "PL", "R", "RUF", "S", "UP", "W"]
ignore = []
fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# https://github.com/astral-sh/ruff/issues/4368
[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = [
    "S101",    #  Use of `assert` detected
    "PLR2004", #  Magic value used in comparison
    "PLR0913", #  Too many arguments in function definition
]
